:pencil2: :: renaming data module -> network module
