+++ core/network/.gitignore
