+++ core/network/proguard-rules.pro
