+++ core/network/src/main/AndroidManifest.xml
